bug fixing

+++ ssd/dataloader.py
@@ -5,7 +5,7 @@ def od_collate_fn(batch):
     labels = []
     imgs = []
     for sample in batch:
-        imgs.append(torch.FloatTensor(sample[0]))
+        imgs.append(sample[0])
         boxes.append(torch.FloatTensor(sample[1]))
         labels.append(torch.FloatTensor(sample[2]))
